21926: New GPU API WIP
Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>

+++ sdk/go/arvados/container.go
@@ -106,7 +106,7 @@ type GPURuntimeConstraints struct {
 	DriverVersion  string   `json:"driver_version"`
 	HardwareTarget []string `json:"hardware_target"`
 	DeviceCount    int      `json:"device_count"`
-	VRAM           int      `json:"vram"`
+	VRAM           int64    `json:"vram"`
 }
 
 // RuntimeConstraints specify a container's compute resources (RAM,
